feat: refactor MLPClassifier and MLPRegressor to use new modularized functions for activation, initialization, and data splitting; make a perceptron classe
