+++ ifri_mini_ml_lib/neural_networks/mlp_regressor.py
@@ -1,5 +1,8 @@
 from typing import List, Tuple, Optional
-from utils import ACTIVATIONS, DERIVATIVES, UPDATE_WEIGHTS_METHODS, TASK_ACTIVATIONS, initialize_weights, split_train_validation
+from optimizers import UPDATE_WEIGHTS_METHODS
+from activation import ACTIVATIONS, DERIVATIVES, TASK_ACTIVATIONS
+from initialization import initialize_weights
+from data_split import split_train_validation
 import numpy as np
 
 
@@ -84,6 +87,8 @@ class MLPRegressor:
         self.early_stopping = early_stopping
         self.validation_fraction = validation_fraction
         self.n_iter_no_change = n_iter_no_change
+        self.random_state = random_state
+        
         
         if random_state is not None:
             np.random.seed(random_state)
@@ -268,7 +273,7 @@ class MLPRegressor:
         
         # Split into training and validation sets if early_stopping is enabled
         if self.early_stopping:
-            X_train, X_val, y_train, y_val = split_train_validation(self, X, y)
+            X_train, X_val, y_train, y_val = split_train_validation( X, y, seed=self.random_state, validation_fraction=self.validation_fraction)
         else:
             X_train, y_train = X, y
